Disable Turbo in `MODE_PARENT` content section (see #7736)
Description
-----------

We do not want Turbo to react to any preview that is output in the back end, e.g. for the parent view of articles where the content elements for the front end are rendered. With https://github.com/contao/contao/pull/7597 this should be a non-issue - however there is a bug in Turbo that prevents Turbo from working properly because of it. But irrespective of this issue it makes sense to disable Turbo for certain back end areas - like the previews in `MODE_PARENT`.

There is two ways to go about it: either add `data-turbo="false"` only to the `<div class="cte_preview">` (which would have to be implemented for each `label_callback` of every DCA) - or we disable it for the whole `tl_content` section of the parent view - and then re-enable Turbo for the operations for example (the `tl_content_right` section).

This PR implements the latter.

Commits
-------

1c647994 Disable Turbo in