Fix prefetching of edit, children and select links (see #7600)
Description
-----------

Replaces https://github.com/contao/contao/pull/7423

I think this is a much safer option than #7423 because we do not try to detect operations, but actually mark the ones we want to prefetch. It also allows others to mark their operations (e.g. see theme modules) as to be prefetched (and therefore not have a request token).

Btw. if the request token would be omitted accidentally on an operation, the warning dialog would be triggered, so that should still be safe.

The new Stimulus controllers fixes the issue of preventing Turbo on the operations due to calling `stopPropagation()` on all links in `Theme.setupCtrlClick`.

Commits
-------

d1fad9d5 Add prefetching of edit, children and select links
eac7b773 Rebuild the assets
09dd3ee7 Correctly handle keyboard keys

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>