Ignore pages in maintenance mode when generating the sitemap (see #7797)
Description
-----------

It seems like the `/sitemap.xml` endpoint contains the page URLs that are under the website root that is in the maintenance mode. This pull request should fix that. Let me know if this is a valid approach to fix this, then I will provide the missing unit tests.

Commits
-------

177eff04 Ignore pages in maintenance mode when generating the sitemap