Use the DBAFS manager in `DC_Folder` (see #7855)
Description
-----------

Currently, `DC_Folder` mostly ignores the VFS/modern DBAFS, which basically prevents us from using the VFS events in the core  (see discussion in #7835).

This PR rewires calls to the old Dbafs implementation to the `DbafsManager`. I only left calls to `Dbafs::shouldBeSynchronized()` if the result of the call was altering the program flow. 

Additionally, when storing `meta` in `tl_files`, we are now also calling `setExtraMetadata()` on the `DbafsManager`, so that subsequent events are triggered.

I'm not really sure if I would consider this a bug or a feature…

/cc @Toflar

Commits
-------

d6ef651d add convenience method to set localized metadata
c30bd3ea use the DbafsManager to synchronize files/directories and to write lo…
6b183e6e update the model cache on sync