Unwrap Twig exceptions (see #7793)
Description
-----------

fixes https://github.com/contao/contao/issues/7783

This unwraps all Twig exceptions supported by our `PrettyErrorScreenListener`. Not sure if that is the correct approach, but also not sure why these exceptions are wrapped in Twig at all?

An alternative could be to unwrap everything in production (e.g. when the pretty error screen is rendered anyway), to correctly allow the `ExceptionConverterListener` and `PrettyErrorScreenListener` to handle the exception.

Commits
-------

375852a2 Unwrap Twig exceptions
2692dd5d Updated unit tests