Fix dynamic Twig inheritance when in a theme context (see #7843)
Description
-----------

Fixes #7836

Basically this is happening right now: our compile-time decisions which template to use from the hierarchy fail in scenarios when there is a theme context (runtime!) and a theme template overwrites/adjusts the template in question. 

With this PR, the compiler only yields a constant name if there are no theme templates for said template, otherwise a match expression is output, that is then evaluated at runtime with the current theme context. 

### Status
- [x] extends tag
- [x] include tag and function
- [x] use tag

Commits
-------

1f9eeff0 fix dynamic inheritance with "extends" for theme contexts
d73a6ac8 fix dynamic inheritance with "use" for theme contexts
39bd6a05 fix dynamic inheritance with "include" for theme contexts
426bfcd7 improve naming and annotations
3a09f196 add tests for ContaoFilesystemLoader