Fix the download mime type (see #7945)
Description
-----------

Fixes `Unable to retrieve metadata from "…": A mime type could not be detected. Set the "$default" argument to suppress this exception.`

This exception gets thrown when trying to download a file in the frontend from the download content element.

Can be reproduced for example with an `.indd` file.

Commits
-------

55e40648 Fix download mime type