Optimize the operations menu (see #8003)
Description
-----------

The `DisclosureMenu` script already handles all necessary options if registered correctly. Now all operations are part of a [Multi-level Disclosure Menu](https://accessible-menu.dev/disclosure-menus.html#multi-level), which also means the submenu closes automatically when navigating outside the DisclosureMenu.

There is one bug currently with using the enter key to activate an operation (see https://github.com/NickDJM/accessible-menu/issues/377), which breaks the toggle handler. But that's unrelated to this PR and was already the case with the previous implementation.

Commits
-------

6fbc0df7 Optimize operations menu