Use Symfony locks instead of DB locks for the `Cron` service (see #8001)
Description
-----------

The `Cron` services tries to get a table lock on `tl_cron_job` when processing which cronjobs to run. This lock is blocking (default is 1 year in MySQL and 1 day in MariaDB), meaning if a second `contao:cron` process is executed, it will wait until the previous is finished. This is usually not an issue as the lock is only held for the duration of reading and writing from and to the `tl_cron_job` table. The lock will be released _before_ the cronjobs are actually executed.

_However_, if something goes awry with the MySQL server, a cascading issue can occur: the `Cron` service might have to wait to get a lock from the table for a long time - or even indefinitely. Then 60 seconds later the next `contao:cron` execution is run - which will also have to wait. Then 60 seconds later the next `contao:cron` execution is run - etc.

This will hog more and more system resources, making matters worse.

This PR fixes that by switching to Symfony Locks instead. @ausi and me also had the idea of introducing a non-block MySQL table lock like this:

```php
public function lockTableNonBlocking(): void
{
    $table = $this->getClassMetadata()->getTableName();

    $defaultLockTimeout = $this->connection->fetchOne("SELECT @@lock_wait_timeout");

    // Use default MySQL lock timeout, if it cannot be retrieved
    if (false === $defaultLockTimeout) {
        $defaultLockTimeout = 31_536_000;
    }

    try {
        // Set a short lock timeout, so that the next statement throws an exception sooner
        $this->connection->executeStatement("SET SESSION lock_wait_timeout = 1");
        $this->connection->executeStatement("LOCK TABLES $table WRITE, $table AS t0 WRITE, $table AS t0_ WRITE");
    } finally {
        // Restore the previous lock timeout
        $this->connection->executeStatement("SET SESSION lock_wait_timeout = ?", [$defaultLockTimeout], [Types::INTEGER]);
    }
}
```

This would then cause Doctrine to throw a `LockWaitTimeoutException` after 1 second, which can be handled in the `Cron` service - instead of blocking indefinitely.

I think it is better to use Symfony Lock instead. The PR uses the default `lock.factory` service from the `symfony/framework-bundle`, which uses by default a `SemaphoreStore` if available, otherwise a `FlockStore`. It can also be customized to another store in the framework bundle config if required.

@ausi prefers the `lockTableNonBlocking()` method as it has the advantage that it actually locks the table that we want to write and read data from. But it could be that this does not work reliably everywhere.

Commits
-------

54aa7b84 Use Symfony locks instead of database locks
87732f95 add deprecation, set default timeout to an hour
440a2999 Merge remote-tracking branch 'origin/5.3' into bugfix/cron-lock
e1d10151 compönent
10517725 compönent
cc7a186a Update the `.draft` styles (see #7951)
fbb7c72d Merge remote-tracking branch 'origin/5.3' into bugfix/cron-lock

Co-authored-by: zoglo <55794780+zoglo@users.noreply.github.com>